pyproject.toml
[project]
name = "graphyard"
version = "0.1.0"
description = "Metrics ingestion and condition evaluation service"
readme = "README.md"
authors = [
    { name = "Jochen Wersdoerfer" }
]
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-environ>=0.12.0",
    "granian>=2.2.6",
    "httpx>=0.28.1",
    "influxdb-client>=1.49.0",
    "whitenoise>=6.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
packages = ["src/graphyard"]

[tool.hatch.build.targets.wheel]
packages = ["src/graphyard"]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "ruff>=0.11.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
addopts = "-ra"
DJANGO_SETTINGS_MODULE = "config.settings.test"
django_find_project = false
pythonpath = [
    "src/django",
    "src",
]

[tool.mypy]
files = ["src/"]
ignore_missing_imports = true
exclude = "src/graphyard/migrations/.*"
warn_unused_ignores = false
warn_return_any = false
disallow_untyped_globals = false
disallow_untyped_defs = false
check_untyped_defs = false

[tool.specsync]
workspace_subdir = "specs"
repo_specs_dir = "specs"

[tool.specsync.filter]
require_expose = true
match_project = true
